functions/requirements.txt
firebase_functions
firebase_admin
google-generativeai
python-dotenv
google-cloud-logging
google-cloud